- ICC compile warning fixes.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@32119 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/genmesh/object/genmesh.h
@@ -413,7 +413,7 @@ public:
   };
   LegacyBuffers legacyBuffers;
   void CreateLegacyBuffers();
-  void ClearLegacyBuffers (uint mask = CS_BUFFER_ALL_MASK);
+  void ClearLegacyBuffers (uint mask = (uint)CS_BUFFER_ALL_MASK);
   void UpdateFromLegacyBuffers();
   
   SubMeshesContainer subMeshes;
